requirements.txt
Numpy
Scipy
Pandas
Matplotlib
Lynx
Makefile
Black
lint
fastapi
uvicorn
python-multipart
pdfplumber
python-docx
pytest
Ipython
aiofiles
pytest-cov
ruff
SpaCy
pymupdf
poppler-utils
pytesseract
pdf2image
Pillow
opencv-python
